Added missing palettes featured in demo page
Based on the way you constructed the .palette-clouds (with a background-color declaration and a separate color declaration), I first just went ahead and took the same approach for all of the color swatches featured in the demo; adding the swatches from the demo that are missing in this partial. However, I realized that if done that way, the results are often too hard to read because there is not enough contrast between the background-color and the color. So I simply added the missing palette background-colors based on what's featured in the great looking demo.  