Made small change to background colour of .first-icon and .second-icon relevant to radio's
Ok this is a very, very small bug I encountered with a project I've been working on. Which involved the setting of white as the background-colour for .radio .icons .second-icon/first-icon at line 1753. Which was resulting in a white box appearing behind the radio buttons on the off-white background of a .well element. It looked a bit tacky so I set the value to inherit but transparent or the removal of the assignment of a background-color would also work.
Enjoy!